AAA/requirements.txt
python-dotenv
requests
python-dateutil
openai
google-genai
tiktoken
SkillsManager
SynLrn
SynMem
BitSig
MediaCapture
